requirements.txt
openai-whisper==20240930
pandas==2.2.3
